Валидацию хэштегов в модуль
